Reduce length of main reducer file
